// src/appConstants/enum.ts
export enum FormField {
  NAME = 'name',
  CONTACT_NUMBER = 'contactNumber',
  EMAIL = 'email',
  PASSWORD = 'password',
  CONFIRM_PASSWORD = 'confirmPassword',
}

export enum ButtonVariant {
  SOLID = 'solid',
  OUTLINE = 'outline',
}

export enum LoginField {
  EMAIL = 'email',
  PASSWORD = 'password',
}